refactor(confidential): relocate amount range validation
