+++ confidential/mptcrypto/errors.go
@@ -1,10 +1,6 @@
 package mptcrypto
 
-import (
-	"errors"
-	"fmt"
-	"math"
-)
+import "errors"
 
 var (
 	// ErrCgoRequired is returned when native confidential MPT operations are unavailable.
@@ -15,13 +11,3 @@ var (
 	// ErrInvalidAmountRange is returned when a decryption search range is invalid.
 	ErrInvalidAmountRange = errors.New("mptcrypto: invalid amount range")
 )
-
-func validateAmountRange(rangeLow, rangeHigh uint64) error {
-	if rangeLow > rangeHigh {
-		return fmt.Errorf("%w: low %d exceeds high %d", ErrInvalidAmountRange, rangeLow, rangeHigh)
-	}
-	if rangeHigh == math.MaxUint64 {
-		return fmt.Errorf("%w: high must be less than %d", ErrInvalidAmountRange, uint64(math.MaxUint64))
-	}
-	return nil
-}
